requirements.txt
aiohttp==3.6.2
discord.py==1.3.4
requests==2.23.0
python-dateutil==2.8.1
